+++ api/src/main/java/org/allaymc/api/world/service/BlockUpdateService.java
@@ -13,13 +13,24 @@ import java.time.Duration;
 public interface BlockUpdateService {
 
     /**
-     * Schedule a block update at the specified position.
-     *
-     * @param pos   The position of the block to update.
-     * @param delay The delay of the block update, in ticks.
+     * @see #scheduleBlockUpdate(Vector3ic, int, int)
      */
     default void scheduleBlockUpdate(Vector3ic pos, int delay) {
         scheduleBlockUpdate(pos, delay, 0);
+    }
+
+    /**
+     * @see #scheduleBlockUpdate(Vector3ic, int, int)
+     */
+    default void scheduleBlockUpdate(Vector3ic pos, Duration delay) {
+        scheduleBlockUpdate(pos, delay, 0);
+    }
+
+    /**
+     * @see #scheduleBlockUpdate(Vector3ic, int, int)
+     */
+    default void scheduleBlockUpdate(Vector3ic pos, Duration delay, int layer) {
+        scheduleBlockUpdate(pos, (int) (delay.toNanos() / 50_000_000), layer);
     }
 
     /**
@@ -32,25 +43,39 @@ public interface BlockUpdateService {
     void scheduleBlockUpdate(Vector3ic pos, int delay, int layer);
 
     /**
-     * Schedule a block update at the specified position.
-     *
-     * @param pos   The position of the block to update
-     * @param delay The delay of the block update
+     * @see #scheduleRandomBlockUpdate(Vector3ic, int, int)
      */
-    default void scheduleBlockUpdate(Vector3ic pos, Duration delay) {
-        scheduleBlockUpdate(pos, delay, 0);
+    default void scheduleRandomBlockUpdate(Vector3ic pos, int delay) {
+        scheduleRandomBlockUpdate(pos, delay, 0);
     }
 
     /**
-     * Schedule a block update at the specified position.
+     * @see #scheduleRandomBlockUpdate(Vector3ic, int, int)
+     */
+    default void scheduleRandomBlockUpdate(Vector3ic pos, Duration delay) {
+        scheduleRandomBlockUpdate(pos, delay, 0);
+    }
+
+    /**
+     * @see #scheduleRandomBlockUpdate(Vector3ic, int, int)
+     */
+    default void scheduleRandomBlockUpdate(Vector3ic pos, Duration delay, int layer) {
+        scheduleRandomBlockUpdate(pos, (int) (delay.toNanos() / 50_000_000), layer);
+    }
+
+    /**
+     * Schedule a random block update at the specified position. Calling this method is considered
+     * equivalent to {@code scheduleBlockUpdate(pos, (int) ((float) delay / (float) randomTickSpeed), layer)}.
+     * <p>
+     * Unlike {@link #scheduleBlockUpdate(Vector3ic, int, int)}, in this method the delay will also
+     * be divided by the random tick speed of the current world. For example, this method is used
+     * by fire, so changing the random tick speed will also change the speed of fire spreading.
      *
-     * @param pos   The position of the block to update
-     * @param delay The delay of the block update
+     * @param pos   The position of the block to update.
+     * @param delay The delay of the block update, in ticks.
      * @param layer The layer of the block update.
      */
-    default void scheduleBlockUpdate(Vector3ic pos, Duration delay, int layer) {
-        scheduleBlockUpdate(pos, (int) (delay.toNanos() / 50_000_000), layer);
-    }
+    void scheduleRandomBlockUpdate(Vector3ic pos, int delay, int layer);
 
     /**
      * Check if a block update is scheduled at the specified position.